ui: consolidate audio data URI handling into shared utility
- Extract getAudioInputFormat to a shared utility (was duplicated inline)
- Store raw base64 in base64Data on the message object
- Use base64Data to construct data URIs for audio rendering
- Update agentic store to build INPUT_AUDIO parts from base64Data
